@@ -7,18 +7,20 @@
     "build": "webpack --mode production"
   },
   "dependencies": {
-    "autoprefixer": "^10.4.0",
+    "@vueuse/core": "^7.5.1",
+    "@vueuse/head": "^0.7.4",
+    "@vueuse/router": "^7.5.1",
     "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.7.2",
     "core-js": "~3.20.1",
-    "vue": "~2.6.14",
-    "vue-filter-pretty-bytes": "~0.1.6",
-    "vue-meta": "~2.4.0",
-    "vue-router": "~3.5.3"
+    "vue": "^3.2.26",
+    "vue-router": "^4.0.12"
   },
   "devDependencies": {
-    "@babel/core": "~7.16.5",
-    "@babel/preset-env": "~7.16.5",
+    "@babel/core": "~7.16.7",
+    "@babel/preset-env": "~7.16.7",
+    "@vue/compiler-sfc": "^3.2.26",
+    "autoprefixer": "^10.4.1",
     "babel-loader": "~8.2.3",
     "compression-webpack-plugin": "~9.2.0",
     "copy-webpack-plugin": "~10.2.0",
@@ -38,7 +40,8 @@
     "mini-css-extract-plugin": "~2.4.5",
     "postcss-html": "^1.3.0",
     "postcss-loader": "~6.2.1",
-    "sass": "~1.45.1",
+    "pretty-bytes": "^5.6.0",
+    "sass": "~1.45.2",
     "sass-loader": "~12.4.0",
     "style-loader": "~3.3.1",
     "stylelint": "~14.2.0",
@@ -46,8 +49,7 @@
     "stylelint-config-standard-scss": "^3.0.0",
     "stylelint-no-unsupported-browser-features": "~5.0.2",
     "svg-inline-loader": "^0.8.2",
-    "vue-loader": "~15",
-    "vue-template-compiler": "~2.6.14",
+    "vue-loader": "^17.0.0",
     "webpack": "~5.65.0",
     "webpack-cli": "~4.9.1",
     "webpack-dev-server": "~4.7.1",
